[Feature #146706545] fix hound errors for infix operators
