+++ package.json
@@ -4,7 +4,7 @@
   "description": "DMS is a full stack document management system, complete with roles and privileges.",
   "main": "index.js",
   "scripts": {
-    "test": "NODE_ENV=test nyc mocha --compilers js:babel-core/register ./**/*.spec.js",
+    "test": "NODE_ENV=test nyc mocha --timeout 15000 --compilers js:babel-core/register ./server/**/*.spec.js",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "test-watch": "nodemon --exec 'npm test'",
     "start": "webpack & babel-node server.js"
