+++ server/tests/routes/users.spec.js
@@ -19,6 +19,7 @@ let documentId = '';
 let roleId = '';
 let adminToken = '';
 
+
 before((done) => {
   Users
     .destroy({ where: {} })
@@ -61,143 +62,6 @@ after((done) => {
             .then(() => done());
         });
     });
-});
-
-
-describe('Document Routes', () => {
-  describe('For Normal Document', () => {
-    describe('POST /documents/', () => {
-      it('should create a new document', (done) => {
-        request.post('/documents')
-          .send({
-            userId: adminId,
-            title: InputDocuments.Public.title,
-            content: InputDocuments.Public.content,
-            access: InputDocuments.Public.access,
-            createdAt: InputDocuments.Public.createdAt,
-            updatedAt: InputDocuments.Public.updatedAt
-          })
-          .set('x-auth', adminToken)
-          .expect(200)
-          .end((err, res) => {
-            documentId = res.body.id;
-            const expected = res.body;
-            const actual = InputDocuments.Public;
-            expect(expected.title).toEqual(actual.title);
-            expect(expected.content).toEqual(actual.content);
-            done(err);
-          });
-      });
-    });
-    describe('POST /documents/', () => {
-      it('should not create for an existing document', (done) => {
-        request.post('/documents')
-          .send({
-            userId: adminId,
-            title: InputDocuments.Public.title,
-            content: InputDocuments.Public.content,
-            access: InputDocuments.Public.access,
-            createdAt: InputDocuments.Public.createdAt,
-            updatedAt: InputDocuments.Public.updatedAt
-          })
-          .set('x-auth', adminToken)
-          .expect(409, done);
-      });
-    });
-
-    describe('GET /documents', () => {
-      it('should get all documents', (done) => {
-        request.get('/documents')
-          .set('x-auth', adminToken)
-          .expect(200)
-          .end((err, res) => {
-            const expected = res.body.length;
-            const actual = 2; // can only get 2 documents
-            expect(expected).toEqual(actual);
-            done(err);
-          });
-      });
-    });
-
-    describe('GET /documents/:id', () => {
-      it('should allow user get his documents', (done) => {
-        request.get(`/documents/${documentId}`)
-          .set('x-auth', adminToken)
-          .expect(200, done);
-      });
-    });
-
-    describe('GET /documents/:id', () => {
-      it('should not get invalid document', (done) => {
-        request.get('/documents/0')
-          .set('x-auth', adminToken)
-          .expect(404, done);
-      });
-    });
-
-    describe('PUT /documents/', () => {
-      it('should be able to update his document details', (done) => {
-        const updatedDetails = {
-          title: 'Updated Name',
-          password: 'updated'
-        };
-        request.put(`/documents/${documentId}`)
-          .set('x-auth', adminToken)
-          .send(updatedDetails)
-          .expect(200)
-          .end((err, res) => {
-            const expected = res.body.msg;
-            const actual = 'Document Updated';
-            expect(expected).toEqual(actual);
-            done(err);
-          });
-      });
-    });
-
-    describe('DELETE /documents/:id', () => {
-      it('should be able to delete document', (done) => {
-        request.delete(`/documents/${documentId}`)
-          .set('x-auth', adminToken)
-          .expect(202)
-          .end((err) => {
-            done(err);
-          });
-      });
-    });
-
-    describe('SEARCH search/users/', () => {
-      it('should be able to search for users', (done) => {
-        request.get('/search/users/?q=admin')
-          .set('x-auth', adminToken)
-          .expect(200)
-          .end((err) => {
-            done(err);
-          });
-      });
-    });
-
-    describe('SEARCH /users/documents', () => {
-      it('should be able to retrieve documents', (done) => {
-        request.get('/search/documents/?q=document')
-          .set('x-auth', adminToken)
-          .expect(200)
-          .end((err) => {
-            done(err);
-          });
-      });
-    });
-
-    describe('SEARCH /users/:id/documents', () => {
-      it('should not search when character is less than 3', (done) => {
-        request.get('/search/documents/?q=do')
-          .set('x-auth', adminToken)
-          .expect(401)
-          .end((err) => {
-            done(err);
-          });
-      });
-    });
-  });
 });
 
 describe('Users Routes', () => {
@@ -438,87 +302,6 @@ describe('Users Routes', () => {
         request.get(`/users/${id}/documents`)
           .set('x-auth', adminToken)
           .expect(200)
-          .end((err) => {
-            done(err);
-          });
-      });
-    });
-  });
-
-  describe('Roles Admin', () => {
-    describe('POST /roles/', () => {
-      it('should create a new role', (done) => {
-        request.post('/roles')
-          .send(InputRoles.NewRole)
-          .set('x-auth', adminToken)
-          .expect(200)
-          .end((err, res) => {
-            roleId = res.body.id;
-            const expected = res.body;
-            const actual = InputRoles.NewRole;
-            expect(expected.title).toEqual(actual.title);
-            done(err);
-          });
-      });
-    });
-
-    describe('POST /roles/', () => {
-      it('should create not create an existing role', (done) => {
-        request.post('/roles')
-          .send(InputRoles.NewRole)
-          .set('x-auth', adminToken)
-          .expect(409, done);
-      });
-    });
-
-
-    describe('GET /roles', () => {
-      it('should be able to get all users', (done) => {
-        request.get('/roles')
-          .set('x-auth', adminToken)
-          .expect(200)
-          .end((err, res) => {
-            const expected = res.body.length;
-            const actual = 3;
-            expect(expected).toEqual(actual);
-            done();
-          });
-      });
-    });
-
-    describe('PUT /roles/', () => {
-      it('should be able to change roles title', (done) => {
-        const updatedDetails = {
-          title: 'Updated Role',
-        };
-        request.put(`/roles/${roleId}`)
-          .set('x-auth', adminToken)
-          .send(updatedDetails)
-          .expect(200)
-          .end((err, res) => {
-            const expected = res.body.msg;
-            const actual = 'Role Updated';
-            expect(expected).toEqual(actual);
-            done(err);
-          });
-      });
-    });
-
-    describe('GET /roles/:id', () => {
-      it('should be able to retrieve roles', (done) => {
-        request.get(`/roles/${roleId}`)
-          .set('x-auth', adminToken)
-          .expect(200)
-          .end((err) => {
-            done(err);
-          });
-      });
-    });
-    describe('DELETE /roles/:id', () => {
-      it('should be able to delete roles', (done) => {
-        request.delete(`/roles/${roleId}`)
-          .set('x-auth', token)
-          .expect(202)
           .end((err) => {
             done(err);
           });
